Refactored viewedProduct unit tests
